change Graph to use 32 bits int

+++ include/Graph.hpp
@@ -12,7 +12,7 @@ concept Arithmetic = std::is_arithmetic_v<T>;
 template <Arithmetic T>
 class Graph {
 	private:
-		uint16_t num_nodes_ = 0;
+		uint32_t num_nodes_ = 0;
 		std::vector< std::vector<T> > paths_weights_; 
 
 	public:
@@ -21,21 +21,21 @@ class Graph {
 		
 		Graph() = default;
 
-		explicit Graph(const uint16_t num_nodes, const T value = 0);
+		explicit Graph(const uint32_t num_nodes, const T value = 0);
 
 
 		// consult methods
-		uint16_t num_nodes() const noexcept;
+		uint32_t num_nodes() const noexcept;
 
-		bool is_connected(const uint16_t first_node, const uint16_t second_node) const noexcept;
+		bool is_connected(const uint32_t first_node, const uint32_t second_node) const noexcept;
 
-		T cost(uint16_t first_node, uint16_t second_node) const noexcept;
+		T cost(uint32_t first_node, uint32_t second_node) const noexcept;
 		
 
 		// paths modification methods
-		void connect(uint16_t first_node, uint16_t second_node, const T cost) noexcept;
+		void connect(uint32_t first_node, uint32_t second_node, const T cost) noexcept;
 
-		void disconnect(const uint16_t first_node, const uint16_t second_node) noexcept;
+		void disconnect(const uint32_t first_node, const uint32_t second_node) noexcept;
